[ram-nand] Fix integer overflows during initialization and I/O
The ram-nand driver accepted user-controlled parameters (page_size,
oob_size, pages_per_block, num_blocks) via CreateDevice without bounds
validation. This could lead to integer overflows when calculating the
device size, causing the driver to map a smaller-than-expected VMO.

Furthermore, read, write, and erase operations used 32-bit arithmetic for
offset calculations, which could wrap around and enable arbitrary memory
access (sandbox escape) in the driver host process.

This change:
1. Adds checks in NandDevice::Init using safemath to ensure that
   page_size, oob_size, pages_per_block, and num_blocks are valid and do
   not overflow.
2. Changes runtime offset, length, and address calculations in
   ReadWriteData, ReadWriteOob, and Erase to use 64-bit math to prevent
   wrapping.
3. Fixes union field assignments in PerformOperation.
4. Adds unit tests verifying that overflow and zero configurations are
   correctly rejected with ZX_ERR_INVALID_ARGS.

Bug: 528400784
Bug: 528400886
Test: fx test ram-nand-test

Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1694694
Commit-Queue: HanBin Yoon <hanbinyoon@google.com>

+++ src/devices/nand/drivers/ram-nand/ram-nand.h
@@ -113,7 +113,9 @@ class NandDevice : public ddk::NandProtocol<NandDevice>,
   void PerformOperations();
   std::pair<bool, std::vector<RamNandOp>> TakePendingOperations();
   void PerformOperation(RamNandOp& operation);
-  uint32_t MainDataSize() const { return params_.NumPages() * params_.page_size; }
+  uint64_t MainDataSize() const {
+    return static_cast<uint64_t>(params_.NumPages()) * params_.page_size;
+  }
 
   // Implementation of the actual commands.
   zx_status_t ReadWriteData(nand_operation_t* operation, bool bytes);
